Camera goes to nearest player
Got rid of "DontGoThroughThings" script
Made it so after 2 seconds, if the puck hasn't hit a player after being
shot, the camera attaches to the nearest player.
This sometimes gives wrong results, because it's based off of the
camera's position, and not the pucks.
In additions, it's calculated by distance, and not like, on the track,
which would probably be better.